Recipe-Organizer-Backend/requirements.txt
Django==1.6.5
Markdown==2.4.1
South==0.8.4
argparse==1.2.1
django-filter==0.7
djangorestframework==2.3.14
wsgiref==0.1.2
